Use `int` for size in utils/hist

+++ src/utils/hist.h
@@ -43,18 +43,18 @@ hist_t;
 
 /* Initializes empty history structure of given size.  Return zero on success,
  * otherwise non-zero is returned. */
-int hist_init(hist_t *hist, size_t size);
+int hist_init(hist_t *hist, int size);
 
 /* Resets content of the history of given size and empties it.  All associated
  * resources are freed. */
-void hist_reset(hist_t *hist, size_t size);
+void hist_reset(hist_t *hist, int size);
 
 /* Checks whether history is empty.  Returns non-zero for empty history,
  * otherwise non-zero is returned. */
 int hist_is_empty(const hist_t *hist);
 
 /* Changes size of the history object from old value to the new one. */
-void hist_resize(hist_t *hist, size_t old_size, size_t new_size);
+void hist_resize(hist_t *hist, int old_size, int new_size);
 
 /* Checks whether given item present in the history.  Returns non-zero if
  * present, otherwise non-zero is returned. */
@@ -64,7 +64,7 @@ int hist_contains(const hist_t *hist, const char item[]);
  * element.  If item is already present in history list, it's moved.  Returns
  * zero when item is added/moved or rejected, on failure non-zero is
  * returned. */
-int hist_add(hist_t *hist, const char item[], size_t size);
+int hist_add(hist_t *hist, const char item[], int size);
 
 #endif /* VIFM__UTILS__HIST_H__ */
